requirements.txt
certifi==2020.6.20
chardet==3.0.4
cursor==1.3.4
idna==2.10
psycopg2-binary==2.8.6
py-postgresql==1.2.2
requests==2.24.0
substitute==1.1.2
urllib3==1.25.10
